+++ marshmallow_jsonapi/__init__.py
@@ -1,8 +1,4 @@
-# -*- coding: utf-8 -*-
 from .schema import Schema, SchemaOpts
 
-__version__ = "0.21.2"
-__author__ = "Steven Loria"
-__license__ = "MIT"
-
+__version__ = "0.22.0"
 __all__ = ("Schema", "SchemaOpts")